create navigation bar, and separate logic into corresponding pages
