+++ frontend/recipe-creator/src/pages/AboutUs.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import './AboutUs.css';
+
+const AboutUs = () => {
+    return (
+        <div className="section">
+            <div className="title">About Us</div>
+            <div className="subtitle">
+                At PantryPal, we believe that cooking should be fun, easy, and accessible to everyone.
+                Our mission is to help you discover new recipes, unleash your creativity in the kitchen, and enjoy delicious meals with your loved ones.
+            </div>
+        </div>
+    );
+};
+
+export default AboutUs;
