+++ frontend/recipe-creator/src/pages/RecipeGenerator.css
@@ -1,0 +1,4 @@
+body {
+    padding: 20px;
+  }
+
