+++ frontend/recipe-creator/src/react-app-env.d.ts
@@ -1,1 +1,5 @@
 /// <reference types="react-scripts" />
+declare module "*.svg" {
+    const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
+    export default content;
+  }
